+++ admin/src/scenes/centersV2/components/modals/ModalTimeSchedule.jsx
@@ -3,7 +3,8 @@ import { Modal } from "reactstrap";
 import CloseSvg from "../../../../assets/Close";
 import { ModalContainer } from "../../../../components/modals/Modal";
 import { BorderButton } from "../../../phase0/components/Buttons";
-import { DeleteButton, DownloadButton } from "../../../phase0/components/commons";
+import { DownloadButton } from "@/scenes/phase0/components/commons/DownloadButton";
+import { DeleteButton } from "@/scenes/phase0/components/commons/DeleteButton";
 import AddImage from "../../../../assets/icons/AddImage";
 import { toastr } from "react-redux-toastr";
 import Loader from "../../../../components/Loader";
