+++ admin/src/scenes/phase0/components/FieldsGroup.jsx
@@ -1,9 +1,9 @@
 import React, { useEffect, useState } from "react";
-import { MiniTitle } from "./commons";
 import CorrectionRequest from "./CorrectionRequest";
 import Field from "./Field";
 import CorrectedRequest from "./CorrectedRequest";
 import DatePickerInput from "@/components/ui/forms/dateForm/DatePickerInput";
+import { MiniTitle } from "./commons/MiniTitle";
 
 export function FieldsGroup({
   name,
